:recycle: Improve code to handle read file error
Create a different function outside main to read contents from the
specified file.

Return a Result type which will return any type of error if there is
problem reading the contents of file and Ok type nothing along with it
if the contents from a file are read successfully.

Also just handle the error type returned from the run function in main.